Removed therapist (leader) from AbstactSession
Change could not tested complete. A lot of tests are still failing.
GroupSessionFactory is fixed caused by this change.
It's possible that views does not work correctly after this change.
But at the moment TherapyView throws an exception caused by incomplete
constructor.
