+++ MhcPms/src/main/java/ch/bfh/red/backend/models/GroupSession.java
@@ -1,18 +1,23 @@
 package ch.bfh.red.backend.models;
 
-import org.hibernate.annotations.Cascade;
-
-import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.ManyToMany;
+
+import org.hibernate.annotations.Cascade;
 
 @Entity
 public class GroupSession extends AbstractSession<GroupSession> {
-	private static final long serialVersionUID = -383961605951531556L;
 
-	@ManyToMany(fetch = FetchType.LAZY)
+	/**
+     * Removed therapist as leader from AbstractSession
+     */
+    private static final long serialVersionUID = 5737573157528543893L;
+
+    @ManyToMany(fetch = FetchType.LAZY)
 	@Cascade({org.hibernate.annotations.CascadeType.MERGE})
 	private List<Patient> patients;
 	
@@ -22,9 +27,9 @@ public class GroupSession extends AbstractSession<GroupSession> {
 
 	public GroupSession() {}
 	
-	public GroupSession(List<Patient> patients, List<Therapist> therapists, Therapist leader, Date startDate, Date endDate,
+	public GroupSession(List<Patient> patients, List<Therapist> therapists, Date startDate, Date endDate,
 			SessionType sessionType) {
-		super(startDate, endDate, sessionType, leader);
+		super(startDate, endDate, sessionType);
 		this.patients = patients;
 		this.therapists = therapists;
 	}
@@ -38,15 +43,7 @@ public class GroupSession extends AbstractSession<GroupSession> {
 	}
 
 	public List<Therapist> getTherapists() {
-		List<Therapist> therapists = new ArrayList<>();
-		for (Therapist therapist : this.therapists) {
-			// iff therapist is the leader, don't want him in this List anymore
-			if (therapist.equals(this.getTherapist())) {
-				continue;
-			}
-			therapists.add(therapist);
-		}
-		return therapists;
+	    return therapists;
 	}
 
 	public void setTherapists(List<Therapist> therapists) {
