+++ MhcPms/src/main/java/ch/bfh/red/backend/models/SingleSession.java
@@ -9,18 +9,27 @@ import javax.persistence.ManyToOne;
 
 @Entity
 public class SingleSession extends AbstractSession<SingleSession> {
-	private static final long serialVersionUID = 8992909023275704028L;
+    
+    /**
+     * Removed therapist as leader from AbstractSession
+     */
+    private static final long serialVersionUID = 8992909023275704028L;
 
-	@ManyToOne
+    @ManyToOne
 	@Cascade({org.hibernate.annotations.CascadeType.MERGE})
 	private Patient patient;
+	
+	@ManyToOne
+    @Cascade({org.hibernate.annotations.CascadeType.MERGE})
+    private Therapist therapist;
 	
 	public SingleSession() {}
 	
 	public SingleSession(Patient patient, Therapist therapist, Date startDate,
 			Date endDate, SessionType sessionType) {
-		super(startDate, endDate, sessionType, therapist);
+		super(startDate, endDate, sessionType);
 		this.patient = patient;
+		this.therapist = therapist;
 	}
 	
 	public Patient getPatient() {
@@ -31,10 +40,19 @@ public class SingleSession extends AbstractSession<SingleSession> {
 		this.patient = patient;
 	}
 	
+	public Therapist getTherapist() {
+        return therapist;
+    }
+
+    public void setTherapist(Therapist therapist) {
+        this.therapist = therapist;
+    }
+	
 	@Override
 	public int hashCode() {
 		final int prime = 31;
 		int result = super.hashCode();
+		result = prime * result + ((therapist == null) ? 0 : therapist.hashCode());
 		result = prime * result + ((patient == null) ? 0 : patient.hashCode());
 		result = prime * result + super.hashCode();
 		return result;
@@ -54,11 +72,11 @@ public class SingleSession extends AbstractSession<SingleSession> {
 				return false;
 		} else if (!patient.equals(other.patient))
 			return false;
-		if (getTherapist() == null) {
-			if (other.getTherapist() != null)
-				return false;
-		} else if (!getTherapist().equals(other.getTherapist()))
-			return false;
+		if (therapist == null) {
+            if (other.therapist != null)
+                return false;
+        } else if (!therapist.equals(other.therapist))
+            return false;
 		return super.equals(obj);
 	}
